TEST/EXAMPLE/REORG: Added c# reload example and tests. Also moved the readkey outside of the run function in pattern reload example so the tests can run properly. In trie/51Degrees.c cast strlen to int16_t to prevent loss of data warning, the length of a property name will never be long enough for this to be a problem.
Related Work Items: #741, #961

+++ src/trie/51Degrees.c
@@ -740,7 +740,7 @@ static fiftyoneDegreesDataSetInitStatus setPropertiesFromExistingDataset(
 		// Count the number of valid properties.
 		requiredProperty = oldDataSet->properties + oldDataSet->requiredProperties[propertyIndex];
 		requiredPropertyName = oldDataSet->strings + requiredProperty->stringOffset;
-		requiredPropertyLength = strlen(requiredPropertyName);
+		requiredPropertyLength = (int16_t)strlen(requiredPropertyName);
 
 		// For each of the available properties in the new dataset:
 		for (index = 0; index < newDataSet->propertiesCount; index++) {
